+++ ProjectApp/Program.cs
@@ -11,6 +11,7 @@ var host = new HostBuilder()
     {
         services.AddSingleton<IProjectDBUtil, DBUtil>();
         services.AddSingleton<IProjectRequestValidator, ProjectRequestValidator>();
+        services.AddSingleton<IProjectAuthorizationService, ProjectAuthorizationService>();
         services.AddSingleton<IProjectService, ProjectService>();
         services.AddHttpClient<IProjectScopeSyncPublisher, ProjectScopeSyncPublisher>();
     })